move user-location components to dedicated directory
